+++ api-gymnasio/src/main/java/com/gymnasio/dto/ClaseRequest.java
@@ -8,7 +8,7 @@ public record ClaseRequest(
         @NotBlank @Size(max = 100) String titulo,
         String descripcion,
         @Size(max = 50) String tipo,
-        @NotNull Long instructorId,
+        @NotNull Integer instructorId,
         LocalDateTime fechaInicio,
         Integer duracionMinutos,
         @NotNull @Min(1) Integer capacidad,
