+++ api-gymnasio/src/main/java/com/gymnasio/service/impl/ClaseServiceImpl.java
@@ -1,7 +1,12 @@
 package com.gymnasio.service.impl;
 
-import com.gymnasio.domain.model.*;
-import com.gymnasio.dto.*;
+import com.gymnasio.domain.model.Clase;
+import com.gymnasio.domain.model.EstadoClase;
+import com.gymnasio.domain.model.EstadoReserva;
+import com.gymnasio.domain.model.Usuario;
+import com.gymnasio.dto.ClaseRequest;
+import com.gymnasio.dto.ClaseResponse;
+import com.gymnasio.dto.DisponibilidadClaseView;
 import com.gymnasio.repository.ClaseRepository;
 import com.gymnasio.repository.UsuarioRepository;
 import com.gymnasio.service.ClaseService;
@@ -26,8 +31,10 @@ public class ClaseServiceImpl implements ClaseService {
 
     @Override
     public ClaseResponse crear(ClaseRequest r) {
+        // r.instructorId() debe ser Integer (no Long)
         Usuario instructor = usuarioRepo.findById(r.instructorId())
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Instructor no existe"));
+
         Clase c = Clase.builder()
                 .titulo(r.titulo())
                 .descripcion(r.descripcion())
@@ -38,25 +45,30 @@ public class ClaseServiceImpl implements ClaseService {
                 .capacidad(r.capacidad())
                 .estado(r.estado() != null ? r.estado() : EstadoClase.PROGRAMADA)
                 .build();
+
         c = claseRepo.save(c);
         return toResponse(c);
     }
 
     @Override
     @Transactional(readOnly = true)
-    public ClaseResponse obtener(Long id) {
-        return claseRepo.findById(id).map(this::toResponse)
+    public ClaseResponse obtener(Integer id) {
+        return claseRepo.findById(id)
+                .map(this::toResponse)
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Clase no encontrada"));
     }
 
     @Override
     @Transactional(readOnly = true)
     public List<ClaseResponse> listar() {
-        return claseRepo.findAll().stream().map(this::toResponse).toList();
+        return claseRepo.findAll()
+                .stream()
+                .map(this::toResponse)
+                .toList();
     }
 
     @Override
-    public ClaseResponse actualizar(Long id, ClaseRequest r) {
+    public ClaseResponse actualizar(Integer id, ClaseRequest r) {
         Clase c = claseRepo.findById(id)
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Clase no encontrada"));
 
@@ -78,11 +90,12 @@ public class ClaseServiceImpl implements ClaseService {
         if (r.estado() != null)
             c.setEstado(r.estado());
 
-        return toResponse(c); // entidad queda administrada, se sincroniza al commit
+        // La entidad está administrada: se sincroniza en el commit
+        return toResponse(c);
     }
 
     @Override
-    public void eliminar(Long id) {
+    public void eliminar(Integer id) {
         if (!claseRepo.existsById(id)) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Clase no encontrada");
         }
@@ -91,19 +104,26 @@ public class ClaseServiceImpl implements ClaseService {
 
     @Override
     @Transactional(readOnly = true)
-    public DisponibilidadClaseView disponibilidad(Long claseId) {
-        // Requiere Reserva + EstadoReserva si quieres el dato real;
-        // si aún no creas Reserva, puedes devolver null o lanzar 501 Not Implemented.
+    public DisponibilidadClaseView disponibilidad(Integer claseId) {
+        // Ajusta el estado si quieres considerar otro filtro
         return claseRepo.disponibilidad(claseId, EstadoReserva.RESERVADA);
     }
 
     private ClaseResponse toResponse(Clase c) {
         String instructorNombre = c.getInstructor().getNombres() + " " + c.getInstructor().getApellidos();
+        // Asegúrate de que ClaseResponse use Integer para los IDs.
         return new ClaseResponse(
-                c.getId(), c.getTitulo(), c.getDescripcion(), c.getTipo(),
-                c.getInstructor().getId(), instructorNombre,
-                c.getFechaInicio(), c.getDuracionMinutos(),
-                c.getCapacidad(), c.getEstado(),
-                c.getFechaCreacion(), c.getFechaActualizacion());
+                c.getId(),
+                c.getTitulo(),
+                c.getDescripcion(),
+                c.getTipo(),
+                c.getInstructor().getId(),
+                instructorNombre,
+                c.getFechaInicio(),
+                c.getDuracionMinutos(),
+                c.getCapacidad(),
+                c.getEstado(),
+                c.getFechaCreacion(),
+                c.getFechaActualizacion());
     }
 }
